aztec: Implement requested changes

+++ core/src/aztec/AZDetectorResult.h
@@ -19,7 +19,7 @@ class DetectorResult : public ZXing::DetectorResult
 	int _nbLayers = 0;
 	bool _readerInit = false;
 	bool _isMirrored = false;
-	uint8_t _runeValue = 0;
+	int _runeValue = 0;
 
 	DetectorResult(const DetectorResult&) = delete;
 	DetectorResult& operator=(const DetectorResult&) = delete;
@@ -29,7 +29,7 @@ public:
 	DetectorResult(DetectorResult&&) noexcept = default;
 	DetectorResult& operator=(DetectorResult&&) noexcept = default;
 
-	DetectorResult(ZXing::DetectorResult&& result, bool isCompact, int nbDatablocks, int nbLayers, bool readerInit, bool isMirrored, uint8_t runeValue)
+	DetectorResult(ZXing::DetectorResult&& result, bool isCompact, int nbDatablocks, int nbLayers, bool readerInit, bool isMirrored, int runeValue)
 		: ZXing::DetectorResult{std::move(result)},
 		  _compact(isCompact),
 		  _nbDatablocks(nbDatablocks),
@@ -46,7 +46,7 @@ public:
 	bool isMirrored() const { return _isMirrored; }
 
 	// Only meaningful is nbDatablocks == 0
-	uint8_t runeValue() const { return _runeValue; }
+	int runeValue() const { return _runeValue; }
 };
 
 } // namespace ZXing::Aztec
